+++ src/FormElement/Checkbox.php
@@ -1,11 +1,6 @@
 <?php
-/**
- * html-form-validator (https://github.com/xtreamwayz/html-form-validator)
- *
- * @see       https://github.com/xtreamwayz/html-form-validator for the canonical source repository
- * @copyright Copyright (c) 2016 Geert Eltink (https://xtreamwayz.com/)
- * @license   https://github.com/xtreamwayz/html-form-validator/blob/master/LICENSE.md MIT
- */
+
+declare(strict_types=1);
 
 namespace Xtreamwayz\HTMLFormValidator\FormElement;
 
@@ -13,7 +8,7 @@ use Zend\Validator\Identical as IdenticalValidator;
 
 class Checkbox extends BaseFormElement
 {
-    protected function getValidators()
+    protected function getValidators() : array
     {
         return [
             [