+++ src/FormElement/Radio.php
@@ -1,11 +1,6 @@
 <?php
-/**
- * html-form-validator (https://github.com/xtreamwayz/html-form-validator)
- *
- * @see       https://github.com/xtreamwayz/html-form-validator for the canonical source repository
- * @copyright Copyright (c) 2016 Geert Eltink (https://xtreamwayz.com/)
- * @license   https://github.com/xtreamwayz/html-form-validator/blob/master/LICENSE.md MIT
- */
+
+declare(strict_types=1);
 
 namespace Xtreamwayz\HTMLFormValidator\FormElement;
 
@@ -13,7 +8,7 @@ use Zend\Validator\InArray as InArrayValidator;
 
 class Radio extends BaseFormElement
 {
-    protected function getValidators()
+    protected function getValidators() : array
     {
         $validators = [];
         $haystack   = [];
@@ -27,9 +22,7 @@ class Radio extends BaseFormElement
 
         $validators[] = [
             'name'    => InArrayValidator::class,
-            'options' => [
-                'haystack' => $haystack,
-            ],
+            'options' => ['haystack' => $haystack],
         ];
 
         return $validators;