+++ src/FormElement/Week.php
@@ -1,11 +1,6 @@
 <?php
-/**
- * html-form-validator (https://github.com/xtreamwayz/html-form-validator)
- *
- * @see       https://github.com/xtreamwayz/html-form-validator for the canonical source repository
- * @copyright Copyright (c) 2016 Geert Eltink (https://xtreamwayz.com/)
- * @license   https://github.com/xtreamwayz/html-form-validator/blob/master/LICENSE.md MIT
- */
+
+declare(strict_types=1);
 
 namespace Xtreamwayz\HTMLFormValidator\FormElement;
 
@@ -13,20 +8,19 @@ use DateInterval;
 use Xtreamwayz\HTMLFormValidator\FormElement\DateTime as DateTimeElement;
 use Zend\Validator\DateStep as DateStepValidator;
 use Zend\Validator\Regex as RegexValidator;
+use function sprintf;
 
 class Week extends DateTimeElement
 {
-    protected function getDateValidator()
+    protected function getDateValidator() : array
     {
         return [
             'name'    => RegexValidator::class,
-            'options' => [
-                'pattern' => '/^[0-9]{4}\-W[0-9]{2}$/',
-            ],
+            'options' => ['pattern' => '/^[0-9]{4}\-W[0-9]{2}$/'],
         ];
     }
 
-    protected function getStepValidator()
+    protected function getStepValidator() : array
     {
         $stepValue = $this->node->getAttribute('step') ?: 1; // Weeks
         $baseValue = $this->node->getAttribute('min') ?: '1970-W01';
@@ -36,7 +30,7 @@ class Week extends DateTimeElement
             'options' => [
                 'format'    => 'Y-\WW',
                 'baseValue' => $baseValue,
-                'step'      => new DateInterval("P{$stepValue}W"),
+                'step'      => new DateInterval(sprintf('P%dW', $stepValue)),
             ],
         ];
     }