+++ test/FormElementsTest.php
@@ -1,22 +1,16 @@
 <?php
-/**
- * html-form-validator (https://github.com/xtreamwayz/html-form-validator)
- *
- * @see       https://github.com/xtreamwayz/html-form-validator for the canonical source repository
- * @copyright Copyright (c) 2016 Geert Eltink (https://xtreamwayz.com/)
- * @license   https://github.com/xtreamwayz/html-form-validator/blob/master/LICENSE.md MIT
- */
+
+declare(strict_types=1);
 
 namespace XtreamwayzTest\HTMLFormValidator;
 
+use PHPUnit\Framework\TestCase;
 use Xtreamwayz\HTMLFormValidator\FormFactory;
 use Xtreamwayz\HTMLFormValidator\ValidationResult;
 
-class FormElementsTest extends \PHPUnit_Framework_TestCase
+class FormElementsTest extends TestCase
 {
-    /**
-     * @dataProvider getIntegrationTests
-     */
+    /** @dataProvider getIntegrationTests */
     public function testIntegration(
         $htmlForm,
         $defaultValues,
@@ -30,7 +24,7 @@ class FormElementsTest extends \PHPUnit_Framework_TestCase
             $this->expectException($expectedException);
         }
 
-        $form   = FormFactory::fromHtml($htmlForm, $defaultValues);
+        $form   = (new FormFactory())->fromHtml($htmlForm, $defaultValues);
         $result = $form->validate($submittedValues);
 
         self::assertInstanceOf(ValidationResult::class, $result);
@@ -214,7 +208,8 @@ class FormElementsTest extends \PHPUnit_Framework_TestCase
 
     private function getDomDocument($html)
     {
-        $doc                     = new \DOMDocument('1.0', 'utf-8');
+        $doc = new \DOMDocument('1.0', 'utf-8');
+
         $doc->preserveWhiteSpace = false;
 
         // Don't add missing doctype, html and body